Implement run_get_pk for running the process as test run (#24)
The singleton pattern for the Runner caused problems when used in an
testing run environment. In this case the single runner instance is
created outside of the trigger. The runner with its event loop is forked
into the trigger environment still referencing the old event loop where
the test run was started. We therefore removed the singleton pattern.

The broker_submit attribue is left as settable in the Runner depending
if it is run or submitted.  Note that the for broker submit in airflow
the api webserver and scheduler need to be running. There is no check
for this at the moment.

Tests and examples have been accordingly updated to use the new function